+++ system_prog/src/system/Bus.java
@@ -19,13 +19,13 @@ public class Bus implements Device{
 	public boolean finish() {
 		return true;
 	}
-	public int load(EDeviceId eDeviceid, int MAR) {
+	public long load(EDeviceId eDeviceid, long MAR) {
 		if(eDeviceid == EDeviceId.eMemory) {
 			return this.memory.load(MAR);
 		}
 		return 0;
 	}
-	public boolean store(EDeviceId eDeviceid, int MAR, int MBR) {
+	public boolean store(EDeviceId eDeviceid, long MAR, long MBR) {
 		if(eDeviceid == EDeviceId.eMemory) {
 //			if (MAR == Memory.SCAN_INTERRUPT) {
 //				System.out.println("enter bus mbr: "+MBR+" "+MAR);
